requirements.txt
discord.py
Flask
python-dotenv
psycopg2-binary
